Module Mirage_crypto

Simpler crypto

Mirage-crypto is a cryptographic library.

The overarching API principle is simply mapping inputs to outputs, wherever feasible.

Similar algorithms in the same class (like hashes or block ciphers) are presented as distinct modules sharing the same signature.

The opam package mirage-crypto-rng provides a cryptographically secure pseudo-random number generator, the package mirage-crypto-pk provides public key cryptography.

Hashing

module Hash : sig ... end

Hashes.

module Poly1305 : sig ... end

The poly1305 message authentication code

Symmetric-key cryptography

module type AEAD = sig ... end

Authenticated encryption with associated data.

module Cipher_block : sig ... end

Block ciphers.

module Chacha20 : sig ... end

The ChaCha20 cipher proposed by D.J. Bernstein.

module Cipher_stream : sig ... end

Streaming ciphers.