Module type S.GCM

Galois/Counter Mode.

include AEAD

val tag_size : int

The size of the authentication tag.

type key

The abstract type for the key.

val of_secret : Cstruct.t -> key

of_secret secret constructs the encryption key corresponding to secret.

raises Invalid_argument
if the length of secret is not a valid key size.

Authenticated encryption and decryption with inline tag

val authenticate_encrypt : key:key -> nonce:Cstruct.t -> ?adata:Cstruct.t -> Cstruct.t -> Cstruct.t

authenticate_encrypt ~key ~nonce ~adata msg encrypts msg with key and nonce, and appends an authentication tag computed over the encrypted msg, using key, nonce, and adata.

raises Invalid_argument
if nonce is not of the right size.

val authenticate_decrypt : key:key -> nonce:Cstruct.t -> ?adata:Cstruct.t -> Cstruct.t -> Cstruct.t option

authenticate_decrypt ~key ~nonce ~adata msg splits msg into encrypted data and authentication tag, computes the authentication tag using key, nonce, and adata, and decrypts the encrypted data. If the authentication tags match, the decrypted data is returned.

raises Invalid_argument
if nonce is not of the right size.

Authenticated encryption and decryption with tag provided separately

val authenticate_encrypt_tag : key:key -> nonce:Cstruct.t -> ?adata:Cstruct.t -> Cstruct.t -> Cstruct.t * Cstruct.t

authenticate_encrypt_tag ~key ~nonce ~adata msg encrypts msg with key and nonce. The computed authentication tag is returned separately as second part of the tuple.

raises Invalid_argument
if nonce is not of the right size.

val authenticate_decrypt_tag : key:key -> nonce:Cstruct.t -> ?adata:Cstruct.t -> tag:Cstruct.t -> Cstruct.t -> Cstruct.t option

authenticate_decrypt ~key ~nonce ~adata ~tag msg computes the authentication tag using key, nonce, and adata, and decrypts the encrypted data. If the authentication tags match, the decrypted data is returned.

raises Invalid_argument
if nonce is not of the right size.

val key_sizes : int array

Key sizes allowed with this cipher.

val block_size : int

The size of a single block.