Module Mirage_crypto.Cipher_block

Block ciphers.

Each algorithm, and each mode of operation, is contained in its own separate module.

module S : sig ... end

Module types for various block cipher modes of operation.

module AES : sig ... end

module DES : sig ... end

val accelerated : [ `XOR | `AES | `GHASH ] list

Operations using non-portable, hardware-dependent implementation in this build of the library.